# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Telemt Panel API
  version: 2.0.0-draft1
  description: >
    Admin API of Telemt Panel v2 plus the unauthenticated subscription page.
    Telemt-shaped payloads follow the upstream source catalog (specs/07-telemt-sdk.md).
servers:
  - url: /
    description: Same origin; base_path is prepended by the SPA runtime when configured.
...
security:
  - session: []
paths:
...
  /api/users:
    get:
      tags: [users]
      operationId: listUsers
      description: >
        Composite of Telemt GET /v1/users merged with quota usage (when the
        `quota` capability is present) and the panel's subscription URL.
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
        "502": { $ref: "#/components/responses/TelemtUnreachable" }
    post:
      tags: [users]
      operationId: createUser
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UserCreate" } } }
      responses:
        "201":
          description: Created; secret returned exactly once
          content:
            application/json:
              schema:
                type: object
                required: [user, secret]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  secret: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: user_exists, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "403": { $ref: "#/components/responses/TelemtReadOnly" }
        "502": { $ref: "#/components/responses/TelemtUnreachable" }
  /api/users/{username}:
    parameters: [ { $ref: "#/components/parameters/Username" } ]
    get:
      tags: [users]
      operationId: getUser
      responses:
        "200": { description: User, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/TelemtUnreachable" }
    patch:
      tags: [users]
      operationId: patchUser
      description: >
        JSON Merge Patch semantics forwarded to Telemt: omitted = unchanged,
        null = remove the limit, value = set.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UserPatch" } } }
      responses:
        "200": { description: Updated user, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "403": { $ref: "#/components/responses/TelemtReadOnly" }
        "502": { $ref: "#/components/responses/TelemtUnreachable" }
    delete:
      tags: [users]
      operationId: deleteUser
      responses:
        "204": { description: Deleted }
        "409": { description: last_user_forbidden, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/TelemtUnreachable" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: panel_session
  parameters:
    Username:
      name: username
      in: path
      required: true
      schema: { type: string, pattern: "^[A-Za-z0-9_.-]{1,64}$" }
  responses:
...
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    BadRequest:
      description: Invalid input
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TelemtUnreachable:
      description: telemt_unreachable | telemt_auth_failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    TelemtReadOnly:
      description: Telemt API is in read_only mode
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - bad_request
            - confirmation_required
            - invalid_credentials
            - rate_limited
            - session_expired
            - csrf_rejected
            - internal_error
            - not_found
            - telemt_unreachable
            - capability_absent
            - capability_unavailable
            - manual_restart_required
            - update_locked
            - sublink_unavailable
            - log_tail_unavailable
            - log_stream_unavailable
            - log_source_error
            - totp_required
            - invalid_totp
            - totp_already_enabled
            - totp_setup_changed
            - invalid_webauthn_origin
            - invalid_webauthn_challenge
            - invalid_webauthn_response
            - webauthn_credential_exists
            - passkey_unavailable
            - telemt_auth_failed
            - user_exists
            - last_user_forbidden
            - read_only
            - revision_conflict
            - invalid_toml
            - invalid_config_path
            - config_unset_unsupported
            - no_changes
            - toml_projection_failed
            - reload_in_progress
            - reload_not_found
            - ambiguous_listeners
            - access_not_editable
            - section_not_editable
            - field_not_editable
            - unauthorized
            - forbidden
            - method_not_allowed
            - config_patch_not_atomic
            - payload_too_large
            - api_disabled
            - maestro_unavailable
            - unsupported_media_type
            - web_runtime_unavailable
            - web_snapshot_busy
            - web_runtime_mismatch
            - web_issuance_enabled
            - web_operation_in_progress
            - web_session_not_found
            - web_operation_not_found
            - web_vhost_not_found
            - web_profile_required
          description: >
            Machine code. Panel codes actually emitted today (grepped from every
            WriteError call site): bad_request, invalid_credentials, rate_limited,
            session_expired, csrf_rejected, internal_error, not_found,
            telemt_unreachable, capability_absent, capability_unavailable,
            manual_restart_required, update_locked, sublink_unavailable,
            log_tail_unavailable, log_stream_unavailable, log_source_error,
            totp_required, invalid_totp, totp_already_enabled,
            totp_setup_changed, invalid_webauthn_origin,
            invalid_webauthn_challenge, invalid_webauthn_response,
            webauthn_credential_exists, passkey_unavailable,
            invalid_toml, invalid_config_path, config_unset_unsupported,
            no_changes, toml_projection_failed.
            capability_absent (501) vs capability_unavailable (503) are
            deliberately distinct, not aliases: capability_absent means the
            route itself doesn't exist on this Telemt build (a bare 404/405
            with no error envelope — detected reactively, after attempting
            the call: rotate-secret, enable/disable, POST /api/telemt/reload,
            GET /api/telemt/reload/{id}); capability_unavailable means the
            route exists but the feature behind it is switched off on this
            Telemt — either known up front from the SDK's cached Capabilities
            probe (GET/PATCH /api/telemt/config, config_api) or reported by
            the response itself (GET /api/telemt/tls-fingerprints, whose
            enabled:false means runtime_edge_enabled is off; read from the
            response rather than probed so an unreachable Telemt still maps
            to 502 telemt_unreachable).
            Reserved for milestones not yet implemented: telemt_auth_failed
            (superseded on /api/telemt/info by a reachable:false
            body, not an error status — kept here for /api/telemt/config, M3).
            A well-formed Telemt *APIError whose status is 4xx and isn't otherwise
            mapped above is passed through verbatim with Telemt's own code — notably
            user_exists, last_user_forbidden, read_only, revision_conflict,
            reload_in_progress, reload_not_found, ambiguous_listeners (the latter two
            absent from Telemt's own documented error-code table but confirmed against
            its source, M3), plus any other code in Telemt's own set (07-telemt-sdk.md):
            bad_request, access_not_editable, section_not_editable, field_not_editable,
            unauthorized, forbidden, method_not_allowed, config_patch_not_atomic,
            payload_too_large, api_disabled, maestro_unavailable — except
            access_not_editable/section_not_editable/field_not_editable/
            config_patch_not_atomic/ambiguous_listeners on PATCH /api/telemt/config,
            which the panel remaps to HTTP 422 regardless of Telemt's own status.
            The WEB group (Telemt >= 3.5.3, internal/telemt/types_web.go) adds
            web_runtime_mismatch, web_issuance_enabled, web_operation_in_progress,
            web_snapshot_busy, web_session_not_found, web_operation_not_found and
            unsupported_media_type; web_runtime_unavailable is listed because it is
            Telemt's own code, but the panel remaps it to capability_unavailable
            (rule R5) so the closed-capability gate is drawn instead of an error.
            Every code in this enum must carry a message in BOTH dictionaries —
            web/src/i18n/i18n.test.ts walks this list.
        message: { type: string }
    User:
      type: object
      description: Telemt UserInfo merged with quota usage and panel extras.
      required: [username, enabled, current_connections, active_unique_ips, total_octets, links]
      properties:
        username: { type: string }
        enabled: { type: boolean }
        in_runtime: { type: boolean }
        user_ad_tag: { type: string }
        max_tcp_conns: { type: integer }
        max_unique_ips: { type: integer }
        data_quota_bytes: { type: integer, format: int64 }
        expiration_rfc3339: { type: string, format: date-time }
        rate_limit_up_bps: { type: integer, format: int64 }
        rate_limit_down_bps: { type: integer, format: int64 }
        current_connections: { type: integer, format: int64 }
        active_unique_ips: { type: integer }
        active_unique_ips_list: { type: array, items: { type: string } }
        recent_unique_ips: { type: integer }
        recent_unique_ips_list: { type: array, items: { type: string } }
        total_octets: { type: integer, format: int64, deprecated: true, description: Raw Telemt process counter retained for compatibility. }
        links: { $ref: "#/components/schemas/UserLinks" }
        quota:
          type: object
          description: Present when the quota capability is available and the user has a quota.
          required: [used_bytes, last_reset_epoch_secs]
          properties:
            used_bytes: { type: integer, format: int64 }
            last_reset_epoch_secs: { type: integer, format: int64 }
        traffic: { $ref: "#/components/schemas/UserTraffic" }
        ip_history: { $ref: "#/components/schemas/UserIPSummary" }
        sub_url: { type: string, description: Absolute subscription page URL }
    UserTraffic:
      type: object
      required: [observed_total_bytes, current_month_bytes, month_key, observed_since_epoch_secs, last_activity_epoch_secs, continuity]
      properties:
        observed_total_bytes: { type: integer, format: int64, minimum: 0 }
        current_month_bytes: { type: integer, format: int64, minimum: 0 }
        month_key: { type: integer, description: UTC month in YYYYMM form }
        observed_since_epoch_secs: { type: integer, format: int64 }
        last_activity_epoch_secs: { type: integer, format: int64 }
        continuity: { $ref: "#/components/schemas/TrafficContinuity" }
    UserLinks:
      type: object
      required: [classic, secure, tls, tls_domains]
      properties:
        classic: { type: array, items: { type: string } }
        secure: { type: array, items: { type: string } }
        tls: { type: array, items: { type: string } }
        tls_domains:
          type: array
          items:
            type: object
            required: [domain, link]
            properties:
              domain: { type: string }
              link: { type: string }
    UserCreate:
      type: object
      required: [username]
      properties:
        username: { type: string, pattern: "^[A-Za-z0-9_.-]{1,64}$" }
        secret: { type: string, pattern: "^[0-9a-fA-F]{32}$" }
        user_ad_tag: { type: string, pattern: "^[0-9a-fA-F]{32}$" }
        max_tcp_conns: { type: integer, minimum: 1 }
        max_unique_ips: { type: integer, minimum: 1 }
        data_quota_bytes: { type: integer, format: int64, minimum: 1 }
        expiration_rfc3339: { type: string, format: date-time }
        rate_limit_up_bps: { type: integer, format: int64, minimum: 1 }
        rate_limit_down_bps: { type: integer, format: int64, minimum: 1 }
        enabled: { type: boolean, default: true }
    UserPatch:
      type: object
      description: JSON Merge Patch — null removes a limit, omitted leaves unchanged.
      properties:
        secret: { type: string, pattern: "^[0-9a-fA-F]{32}$" }
        user_ad_tag: { type: [string, "null"] }
        max_tcp_conns: { type: [integer, "null"] }
        max_unique_ips: { type: [integer, "null"] }
        data_quota_bytes: { type: [integer, "null"], format: int64 }
        expiration_rfc3339: { type: [string, "null"] }
        rate_limit_up_bps: { type: [integer, "null"], format: int64 }
        rate_limit_down_bps: { type: [integer, "null"], format: int64 }
        enabled: { type: [boolean, "null"] }
...
    TrafficContinuity:
      type: string
      enum: [normal, partial]
...
    UserIPSummary:
      type: object
      required: [unique, from, collected_through, history_limited, collection_gap]
      properties:
        unique: { type: integer, format: int64, minimum: 0 }
        from: { type: integer, format: int64, description: Last 30 days, clipped by effective retention. }
        collected_through: { type: integer, format: int64 }
        history_limited: { type: boolean }
        collection_gap: { type: boolean }
...